arch: E-series — 3 more route groups, codec_dashboard 2921→2530
Continuation of dashboard route extraction (PR #162 follow-up).
Pytest: 2172 → 2185 (+13 regression tests, 0 regressions).

EXTRACTIONS

E1 / SR-47: Sparkle update endpoints → routes/update.py
  2 endpoints (/api/update/check, /api/update/download). Both proxy
  through codec_update which handles Ed25519 signature verification.

E2 / SR-48: public health + manifest + metrics + status → routes/health.py
  5 endpoints (/api/health, /health, /api/status, /manifest.json, /metrics).
  HealthResponse Pydantic model moved with them. /api/services/status
  stays in codec_dashboard.py because it reads _bg_status / _bg_tasks
  globals owned by the startup hooks.

E4 / SR-49: upload + upload_image + save_file → routes/upload.py
  3 security-hardened endpoints bundled because they share defensive
  scaffolding:
    - _UPLOAD_MAX_BYTES + 3-layer size cap (B1 / SR-15)
    - _fence_user_document marker helper (B1 / SR-16)
    - _SAVE_FILE_BLOCKED_* + _save_file_is_safe (A1 / SR-8 mirror of
      PR-1C file_write blocklist)

DEFERRED (next round)
  - E3 routes/config.py — /api/config GET/PUT with 22-rule validation
    matrix + sensitive field masking. Medium-risk; standalone PR.
  - E5 routes/vibe_exec.py — /api/preview, /preview_frame, /api/run_code
    (Vibe IDE exec surface; touches subprocess + is_dangerous check)
  - E6 routes/misc.py — /api/tts, /api/response, /api/web_search,
    /api/command, /api/cdp/status

LOC SHIFT

  codec_dashboard.py: 2,921 → 2,530 LOC (-391, -13.4%)
                     (3,912 → 2,530 cumulative = -1,382, -35.3%)

UPDATED TESTS

- tests/test_top10_action_list.py: `_save_file_is_safe` import points
  to routes.upload (was codec_dashboard).
- tests/test_a12_invariant.py: routes/upload.py allowlisted for vision
  POST (A-11 pending, same as routes/media.py).
- tests/test_dashboard_llm.py: /chat/completions count assertion
  bumped from 4 → 3 (upload_image vision POST moved out).
- tests/test_dead_code_3g.py: HealthResponse now lives at
  routes.health.HealthResponse (E2 extraction).
- tests/test_route_extractions_e.py NEW: 13 tests parametrized across
  3 phases — endpoint-registered + source-side invariants
  (HealthResponse exported, B1 helpers in routes/upload).

NEW INVARIANT FLOOR

  test_dashboard_loc_below_2700 — codec_dashboard.py < 2700 LOC.
  Tightens as future extractions land.

Full pytest: 2185 passed, 83 skipped (env-dependent), 0 failed (1m20s).
